requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_pytorch_select=0.2=gpu_0
_tflow_select=2.3.0=mkl
absl-py=0.9.0=py37_0
alabaster=0.7.12=py37_0
asn1crypto=1.3.0=py37_0
astroid=2.3.3=py37_0
astunparse=1.6.3=py_0
attrs=19.3.0=py_0
babel=2.8.0=py_0
backcall=0.1.0=py37_0
blas=1.0=mkl
bleach=3.1.0=py37_0
blinker=1.4=py37_0
bokeh=1.4.0=py37_0
c-ares=1.15.0=h7b6447c_1001
ca-certificates=2020.7.22=0
cachetools=4.1.1=py_0
certifi=2020.6.20=py37_0
cffi=1.14.0=py37h2e261b9_0
chardet=3.0.4=py37_1003
click=7.1.2=py_0
cloudpickle=1.3.0=py_0
cryptography=2.8=py37h1ba5d50_0
cudatoolkit=10.0.130=0
cudnn=7.6.5=cuda10.0_0
cycler=0.10.0=py37_0
cytoolz=0.10.1=py37h7b6447c_0
dask-core=2.25.0=py_0
dbus=1.13.12=h746ee38_0
decorator=4.4.1=py_0
defusedxml=0.6.0=py_0
docutils=0.16=py37_0
entrypoints=0.3=py37_0
expat=2.2.6=he6710b0_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
gast=0.3.3=py_0
glib=2.63.1=h5a9c865_0
gmp=6.1.2=h6c8ec71_1
google-auth=1.21.1=py_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.2.0=py_0
grpcio=1.31.0=py37hf8bcb03_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.10.0=py37hd6299e0_1
hdf5=1.10.6=hb1b8bf9_0
icu=58.2=h9c2bf20_1
idna=2.8=py37_0
imageio=2.9.0=py_0
imagesize=1.2.0=py_0
importlib-metadata=1.7.0=py37_0
importlib_metadata=1.7.0=0
intel-openmp=2020.0=166
ipykernel=5.1.4=py37h39e3cac_0
ipython=7.12.0=py37h5ca1d4c_0
ipython_genutils=0.2.0=py37_0
ipywidgets=7.5.1=py_0
isort=4.3.21=py37_0
jedi=0.16.0=py37_0
jeepney=0.4.2=py_0
jinja2=2.11.1=py_0
joblib=0.16.0=py_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=py37_0
jupyter=1.0.0=py37_7
jupyter_client=5.3.4=py37_0
jupyter_console=6.1.0=py_0
jupyter_core=4.6.1=py37_0
keras=2.4.3=0
keras-base=2.4.3=py_0
keras-preprocessing=1.1.0=py_1
keyring=21.1.0=py37_0
kiwisolver=1.1.0=py37he6710b0_0
lazy-object-proxy=1.4.3=py37h7b6447c_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.13.0=hd408876_0
libsodium=1.0.16=h1bed415_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
markdown=3.2.2=py37_0
markupsafe=1.1.1=py37h7b6447c_0
matplotlib=3.1.3=py37_0
matplotlib-base=3.1.3=py37hef1b27d_0
mccabe=0.6.1=py37_1
mistune=0.8.4=py37h7b6447c_0
mkl=2020.0=166
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.0.15=py37ha843d7b_0
mkl_random=1.1.0=py37hd6b4f25_0
nbconvert=5.6.1=py37_0
nbformat=5.0.4=py_0
ncurses=6.2=he6710b0_0
networkx=2.5=py_0
nibabel=3.0.0=pypi_0
ninja=1.9.0=py37hfd86e86_0
notebook=6.0.3=py37_0
numpy=1.18.1=py37h4f9e942_0
numpy-base=1.18.1=py37hde5b4d6_1
numpydoc=0.9.2=py_0
oauthlib=3.1.0=py_0
olefile=0.46=py37_0
openssl=1.1.1g=h7b6447c_0
opt_einsum=3.1.0=py_0
packaging=20.1=py_0
pandas=1.0.1=py37h0573a6f_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py37_1
parso=0.6.1=py_0
patsy=0.5.1=py37_0
pcre=8.43=he6710b0_0
pexpect=4.8.0=py37_0
pickleshare=0.7.5=py37_0
pillow=6.2.2=pypi_0
pip=20.0.2=py37_1
prometheus_client=0.7.1=py_0
prompt_toolkit=3.0.3=py_0
protobuf=3.13.0=py37hf484d3e_0
psutil=5.6.7=py37h7b6447c_0
ptyprocess=0.6.0=py37_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycodestyle=2.5.0=py37_0
pycparser=2.19=py37_0
pyflakes=2.1.1=py37_0
pygments=2.5.2=py_0
pyjwt=1.7.1=py37_0
pylint=2.4.4=py37_0
pymatreader=0.0.23=pypi_0
pyopenssl=19.1.0=py37_0
pyparsing=2.4.6=py_0
pyqt=5.9.2=py37h05f1152_2
pyrsistent=0.15.7=py37h7b6447c_0
pysocks=1.7.1=py37_0
python=3.7.6=h0371630_2
python-dateutil=2.8.1=py_0
pytorch=1.3.1=cuda100py37h53c1284_0
pytorch-gpu=1.3.1=0
pytz=2019.3=py_0
pywavelets=1.1.1=py37h7b6447c_0
pyyaml=5.3=py37h7b6447c_0
pyzmq=18.1.1=py37he6710b0_0
qt=5.9.7=h5867ecd_1
qtawesome=0.6.1=py_0
qtconsole=4.6.0=py_1
qtpy=1.9.0=py_0
readline=7.0=h7b6447c_5
requests=2.22.0=py37_1
requests-oauthlib=1.3.0=py_0
rope=0.16.0=py_0
rsa=4.6=py_0
scikit-image=0.16.2=py37h0573a6f_0
scikit-learn=0.22.1=py37hd81dba3_0
scipy=1.4.1=py37h0b6359f_0
seaborn=0.10.0=py_0
secretstorage=3.1.2=py37_0
send2trash=1.5.0=py37_0
setuptools=45.2.0=py37_0
sip=4.19.8=py37hf484d3e_0
six=1.14.0=py37_0
snowballstemmer=2.0.0=py_0
sphinx=2.4.0=py_0
sphinxcontrib=1.0=py37_1
sphinxcontrib-applehelp=1.0.2=py_0
sphinxcontrib-devhelp=1.0.2=py_0
sphinxcontrib-htmlhelp=1.0.3=py_0
sphinxcontrib-jsmath=1.0.1=py_0
sphinxcontrib-qthelp=1.0.3=py_0
sphinxcontrib-serializinghtml=1.1.4=py_0
sphinxcontrib-websupport=1.2.0=py_0
spyder=3.3.6=py37_0
spyder-kernels=0.5.2=py37_0
sqlite=3.31.1=h7b6447c_0
statsmodels=0.11.0=py37h7b6447c_0
tensorboard=2.2.1=pyh532a8cf_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.2.0=mkl_py37h6e9ce2d_0
tensorflow-base=2.2.0=mkl_py37hd506778_0
tensorflow-estimator=2.2.0=pyh208ff02_0
termcolor=1.1.0=py37_1
terminado=0.8.3=py37_0
testpath=0.4.4=py_0
tk=8.6.8=hbc83047_0
toolz=0.10.0=py_0
torchvision=0.4.2=cuda100py37hecfc37a_0
tornado=6.0.3=py37h7b6447c_3
traitlets=4.3.3=py37_0
urllib3=1.25.8=py37_0
wcwidth=0.1.8=py_0
webencodings=0.5.1=py37_1
werkzeug=1.0.1=py_0
wheel=0.34.2=py37_0
widgetsnbextension=3.5.1=py37_0
wrapt=1.11.2=py37h7b6447c_0
wurlitzer=2.0.0=py37_0
xlrd=1.2.0=py37_0
xmltodict=0.12.0=pypi_0
xz=5.2.4=h14c3975_4
yaml=0.1.7=had09818_2
zeromq=4.3.1=he6710b0_3
zipp=3.1.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
